third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.recursive.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins cannot open their own popin
  assert_equals: expected "Could not open inner popin" but got "Did open inner popin"
Harness: the test ran to completion.
